add productRestController

+++ product/src/main/java/com/vnco/fusiontech/product/service/ProductService.java
@@ -10,8 +10,8 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getAllProducts();
-    Product getProductById(Long id);
+    Product getProductById(int id);
     Product createProduct(Product product);
     Product updateProduct(Product product);
-    void deleteProduct(Long id);
+    void deleteProduct(int id);
 }
